# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Healthtech Scheduling API",
        "description": "Agendamento de consultas: pacientes, profissionais e ciclo de vida da consulta.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Servidor atual"
        }
    ],
    "paths": {
...
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "description": "Autentica com e-mail e senha e retorna um token Sanctum (bearer). Resposta idêntica para senha errada ou e-mail inexistente (anti-enumeração).",
                "operationId": "985cd5203e2d0a820e113f2f4fdb2847",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "staff@dragenda.test"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login bem-sucedido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Credenciais inválidas ou dados ausentes"
                    },
                    "429": {
                        "description": "Muitas tentativas — aguarde e tente de novo"
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout",
                "description": "Revoga o token atual.",
                "operationId": "a05022a34eef599e04fb0a5254375b84",
                "responses": {
                    "204": {
                        "description": "Token revogado"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Usuário autenticado",
                "operationId": "e339e0e74860c0647fd1d41bbbbc13b2",
                "responses": {
                    "200": {
                        "description": "Dados do usuário autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/patients": {
            "get": {
                "tags": [
                    "Patients"
                ],
                "summary": "Lista pacientes",
                "operationId": "25da496d74a0ef65407930b3a958b687",
                "responses": {
                    "200": {
                        "description": "Lista paginada de pacientes"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Sem permissão (só staff)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Patients"
                ],
                "summary": "Cadastra um paciente",
                "operationId": "1c5763b2aa6fe7d167f31b0a017a4ee6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "cpf",
                                    "phone",
                                    "birth_date"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Maria Silva"
                                    },
                                    "cpf": {
                                        "type": "string",
                                        "example": "111.444.777-35"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "11999998888"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "maria@example.com",
                                        "nullable": true
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-05-10"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Paciente criado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Patient"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação (CPF inválido/duplicado, etc.)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/patients/{patient}": {
            "get": {
                "tags": [
                    "Patients"
                ],
                "summary": "Detalhe de um paciente",
                "operationId": "c11edbbc003c976c0403d7de42a041da",
                "parameters": [
                    {
                        "name": "patient",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paciente"
                    },
                    "404": {
                        "description": "Não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Patients"
                ],
                "summary": "Atualiza um paciente",
                "operationId": "5a4810febcaedc2b35e3594d2b71c457",
                "parameters": [
                    {
                        "name": "patient",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paciente atualizado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Patient": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Maria Silva"
                    },
                    "cpf": {
                        "type": "string",
                        "example": "11144477735"
                    },
                    "phone": {
                        "type": "string",
                        "example": "11999998888"
                    },
                    "email": {
                        "type": "string",
                        "example": "maria@example.com",
                        "nullable": true
                    },
                    "birth_date": {
                        "type": "string",
                        "format": "date",
                        "example": "1990-05-10"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Recepção Central"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "staff@dragenda.test"
                    },
                    "role": {
                        "type": "string",
                        "example": "staff",
                        "enum": [
                            "staff",
                            "patient"
                        ]
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token obtido em POST /auth/login. Enviar como \"Authorization: Bearer {token}\".",
                "bearerFormat": "Sanctum token",
                "scheme": "bearer"
            }
        }
    },
...
}